Move sign configuration classes to new package Add sign create entry command
