+++ cloudnet-modules/cloudnet-signs/src/main/java/de/dytanic/cloudnet/ext/signs/CloudNetSignListener.java
@@ -3,6 +3,8 @@ package de.dytanic.cloudnet.ext.signs;
 import de.dytanic.cloudnet.driver.event.EventListener;
 import de.dytanic.cloudnet.driver.event.events.channel.ChannelMessageReceiveEvent;
 import de.dytanic.cloudnet.driver.event.events.service.*;
+import de.dytanic.cloudnet.ext.signs.configuration.SignConfiguration;
+import de.dytanic.cloudnet.ext.signs.configuration.SignConfigurationProvider;
 
 public final class CloudNetSignListener {
