+++ cloudnet-modules/cloudnet-signs/src/main/java/de/dytanic/cloudnet/ext/signs/bukkit/listener/BukkitSignInteractionListener.java
@@ -3,9 +3,9 @@ package de.dytanic.cloudnet.ext.signs.bukkit.listener;
 import de.dytanic.cloudnet.ext.bridge.BridgePlayerManager;
 import de.dytanic.cloudnet.ext.signs.AbstractSignManagement;
 import de.dytanic.cloudnet.ext.signs.Sign;
-import de.dytanic.cloudnet.ext.signs.SignConfigurationEntry;
-import de.dytanic.cloudnet.ext.signs.SignConfigurationProvider;
 import de.dytanic.cloudnet.ext.signs.bukkit.BukkitSignManagement;
+import de.dytanic.cloudnet.ext.signs.configuration.SignConfigurationProvider;
+import de.dytanic.cloudnet.ext.signs.configuration.entry.SignConfigurationEntry;
 import org.bukkit.ChatColor;
 import org.bukkit.Location;
 import org.bukkit.event.EventHandler;
