+++ cloudnet-modules/cloudnet-signs/src/main/java/de/dytanic/cloudnet/ext/signs/nukkit/command/CommandCloudSign.java
@@ -9,7 +9,11 @@ import cn.nukkit.command.Command;
 import cn.nukkit.command.CommandSender;
 import cn.nukkit.level.Location;
 import de.dytanic.cloudnet.common.collection.Iterables;
-import de.dytanic.cloudnet.ext.signs.*;
+import de.dytanic.cloudnet.ext.signs.AbstractSignManagement;
+import de.dytanic.cloudnet.ext.signs.Sign;
+import de.dytanic.cloudnet.ext.signs.SignPosition;
+import de.dytanic.cloudnet.ext.signs.configuration.SignConfigurationProvider;
+import de.dytanic.cloudnet.ext.signs.configuration.entry.SignConfigurationEntry;
 import de.dytanic.cloudnet.ext.signs.nukkit.NukkitSignManagement;
 import de.dytanic.cloudnet.wrapper.Wrapper;
