+++ cloudnet/src/main/resources/lang/english.properties
@@ -248,6 +248,7 @@ module-smart-stop-service-automatically=CloudService [uniqueId=%id% task=%task% 
 # Module Signs
 module-signs-command-signs-description=Simple administration of the Signs module
 module-signs-command-reload-success=The signs configuration has been reloaded
+module-signs-command-create-entry-success=The sign configuration entry has been created
 # Module Cloudflare
 module-cloudflare-create-dns-record-for-service=DNS Entry on domain %domain% was created for service %service% with Id %recordId%
 module-cloudflare-delete-dns-record-for-service=DNS entry was deleted on domain %domain% for service %service% with Id %recordId%
