+++ cloudnet/src/main/resources/lang/german.properties
@@ -248,6 +248,7 @@ module-smart-stop-service-automatically=CloudService [uniqueId=%id% task=%task% 
 # Module Signs
 module-signs-command-signs-description=Einfache Verwaltung des Signs-Moduls
 module-signs-command-reload-success=Die Schilder Konfiguration wurde erfolgreich neu geladen
+module-signs-command-create-entry-success=Der Konfigurationseintrag wurde erfolgreich hinzugefügt
 # Module Cloudflare
 module-cloudflare-create-dns-record-for-service=DNS Eintrag auf die Domain %domain% wurde für den Service %service% mit der ID %recordId% erstellt
 module-cloudflare-delete-dns-record-for-service=DNS Eintrag auf der Domain %domain% wurde für den Service %service% mit der ID %recordId% gelöscht
